Move 'TelegramMessageManager' into the 'services.telegram' namespace
